Update:
Now the program checks for site status using playwright instead of requests because it's more accurate in detecting site status.

Updated a few warning messages for better user interaction.

Added new config param 'iscaptcha' so the program or the user manually can set it to 'true' or 'false'

When 'iscaptcha' is enabled the script runs with headless mode turned OFF until it cools down.
*fuck off google captcha*
....